[fragment] Add exponential backoff support
When downloading a HLS video I noted that many fragments were missing.
Looking at the log, these fragments were given transient 429s in order
to suggest to the client that it should back off and try again after a
few seconds. However, the current retries happen as fast as we can go,
which means those fragments just end up skipped.

This patch adds support for exponential backoff when doing fragment
retries, which increases the chance that the fragment will eventually be
provided in such cases. This is disabled by default, and can be enabled
with --fragment-retry-backoff.

Example delays in seconds are as follows with the default 10 retries:

    >>> for i in range(10): print("{:.2f}".format(0.25 * 1.5 ** i))
    0.25
    0.38
    0.56
    0.84
    1.27
    1.90
    2.85
    4.27
    6.41
    9.61